Update for the end of the night:
comms contains two feature that I'm working on. The first is a movement
system that makes movements based on enemy robot locations. I have it
setup for scouts, so that they stay out of sight range of enemies while
scouting (if possible). There is also the start of a comms system (hence
the name). That's more complicated and a major WIP, so don't worry too
much about that yet. 

Also, I made a map for testing.