Working on advanced scout movement:
Currently been working on advanced scout movement and tasks. Scouts will
now prioritize seeking out and shaking bullet trees. If there are no
trees, they will try to do combat with enemies. If there are no enemies,
they will just scout, while staying spread out. 

Tree shaking methods are all done. Might be some small
changes/optimization, but not much left on this. 

Combat related stuff is a heavy work in progress. I've been testing a
few things, will be adding back in some old things, and will have a good
deal to adjust. 

General motion is mostly panned out. There are a few things to clean up,
but functionally I think it is all there. 

For all the movement stuff (including a modified tryMove method), I am
returning and tracking the direction of motion so the scouts can use
that to act more logically. This is something I'd like to add to all of
the robot classes. 

Finally, there are methods for searching the trees for robots and
logging that information for other robots to process. I'm not actually
implementing this yet, as it is crazy expensive, and it might be better
suited on another robot type.