First big chunk of making moving smarter
New movement philosophy: update intended move direction based on
different conditions and move at the end. Some conditions might override
the standard flow (being attacked, needing to attack, etc), but there
will be a standard flow of movement for when things aren't happening.
This will include checking friendly unit separation, dodging, and being
able to move. Will improve this over time.

Took the scout unit separation feature and added to all classes. Will
definitely need to tune the separation variables, as well as some of the
other behaviors, but it's a start.  

Updated bullet dodging to the new system, with a slightly different
algorithm (similar to separation) that accounts for all bullets.
Eventually want to weight the dodging by distance from bullet and bullet
damage potential. 

Updated tryMove so that it passes a direction instead of boolean to
track where we actually end up moving and use that as the base for the
next move. 

Added a bunch of separation constants

Added some formatting and commenting