Big Update!!!!!
Okay, this is the big one. We now have much much smarter Archons and
Gardeners. There is still a lot of room for improvement but the backbone
is there. 

Gardeners now make a grid based grove, and Archons patrol them. They
signal if we need lumberjacks, to defend, or to attack. 

Also, removed a lot of the old players.